[modules/fru] Use size_t variable for FRU access functions
We need a larger type for the variable that specifies the length of the
buffer to read/written given that the memories can be larger than 256
bytes

+++ modules/fru.h
@@ -35,8 +35,8 @@ enum {
 };
 
 typedef size_t (* fru_build_t)(uint8_t **buffer);
-typedef uint8_t (* fru_st_read_t)(uint8_t id, uint16_t address, uint8_t *buffer, uint8_t len, uint32_t timeout);
-typedef uint8_t (* fru_st_write_t)(uint8_t id, uint16_t address, uint8_t *buffer, uint8_t len, uint32_t timeout);
+typedef size_t (* fru_st_read_t)(uint8_t id, uint16_t address, uint8_t *buffer, size_t len, uint32_t timeout);
+typedef size_t (* fru_st_write_t)(uint8_t id, uint16_t address, uint8_t *buffer, size_t len, uint32_t timeout);
 
 typedef struct fru_cfg {
     uint8_t eeprom_id;
